pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "blazel"
authors = [
    { name = "Christian Schäfer", email = "christian.schaefer@tatenmitdaten.com" },
]
description = "Package for running extract load jobs on AWS Lambda"
requires-python = ">=3.12"
dynamic = ["version"]

dependencies = [
    "snowflake-connector-python",
    "pyyaml",
    "typer",
    "boto3-stubs[s3,dynamodb,secretsmanager,lambda,stepfunctions]",
]

[tool.setuptools.packages.find]
where = ["."]

[project.optional-dependencies]
dev = [
    "boto3",
    "pytest",
    "moto[s3,dynamodb]",
    "types-PyYAML",
    "mypy",
    "flake8",
]
parquet = [
    "pyarrow"
]

[project.scripts]
blazel = "blazel.cli:cli"